rename classes.py to platforms.py
