+++ tests/performance_tests/cprofile_test.py
@@ -19,7 +19,7 @@ import sys
 
 from count_files.utils.viewing_modes import show_result_for_search_files, show_2columns, show_result_for_total
 from count_files.__main__ import main_flow
-from count_files.classes import get_current_os
+from count_files.platforms import get_current_os
 
 
 def get_locations(*args):
